Extract checkout confirm button
Adds error state for failed tx and wrong network
Adds onConfirmed logic
