+++ packages/app/provider/index.tsx
@@ -45,4 +45,5 @@ const Providers = compose([
   SafeAreaProvider,
   TamaguiProvider,
   ToastProvider,
+  QueryClientProvider,
 ])
